Rename files that refer to plugins
Previously we called plugins 'extensions'. Now that we have decided on
a new name, we should make sure our code uses the new name where
appropriate.

This commit renames all files that have 'extensions' in the
name so that they now have 'plugins' in the name.
