+++ .gitignore
@@ -13,7 +13,7 @@ usage-data-config.json
 /cypress/screenshots
 /cypress/temp
 /cypress/videos
-/cypress/fixtures/extensions/*/package-lock.json
+/cypress/fixtures/plugins/*/package-lock.json
 /govuk-prototype-kit*.zip
 /govuk-prototype-kit*.tgz
 tmp/
