+++ cypress/e2e/5-plugin-tests/install-plugin-test.cypress.js
