+++ cypress/e2e/5-plugin-tests/single-plugin-test.cypress.js
