+++ cypress/fixtures/plugins/plugin-bar/filters.js
