+++ cypress/fixtures/plugins/plugin-bar/package.json
