+++ cypress/fixtures/plugins/plugin-bar/scripts/bar.js
