+++ cypress/fixtures/plugins/plugin-bar/views/bar.njk
