+++ cypress/fixtures/plugins/plugin-baz/package.json
