+++ cypress/fixtures/plugins/plugin-baz/sass/baz.scss
