+++ cypress/fixtures/plugins/plugin-baz/scripts/baz.js
