+++ cypress/fixtures/plugins/plugin-baz/views/baz.njk
