+++ cypress/fixtures/plugins/plugin-foo/filters.js
