+++ cypress/fixtures/plugins/plugin-foo/package.json
