+++ cypress/fixtures/plugins/plugin-foo/sass/foo.scss
