+++ cypress/fixtures/plugins/plugin-foo/scripts/foo.js
