+++ cypress/fixtures/plugins/plugin-foo/views/foo.njk
