+++ cypress/scripts/run-starter-prototype.js
@@ -10,8 +10,8 @@ const testDir = path.resolve(process.env.KIT_TEST_DIR || defaultKitPath)
 ;(async () => {
   await mkPrototype(testDir, { overwrite: true, allowTracking: false, npmInstallLinks: true })
 
-  const fooLocation = path.join(__dirname, '..', 'fixtures', 'extensions', 'extension-foo')
-  const barLocation = path.join(__dirname, '..', 'fixtures', 'extensions', 'extension-bar')
+  const fooLocation = path.join(__dirname, '..', 'fixtures', 'plugins', 'plugin-foo')
+  const barLocation = path.join(__dirname, '..', 'fixtures', 'plugins', 'plugin-bar')
 
   await installExtensions(testDir, [
     '@govuk-prototype-kit/step-by-step@1',
