+++ lib/nunjucks/govuk-prototype-kit/includes/stylesheets.html
@@ -1,3 +1,3 @@
 <link href="/public/stylesheets/application.css" rel="stylesheet" type="text/css" />
 
-{% include "./stylesheets-extensions.njk" %}
+{% include "./stylesheets-plugins.njk" %}
