+++ lib/plugins/plugins.js
