+++ lib/plugins/plugins.test.js
@@ -5,7 +5,7 @@ const path = require('path')
 const config = require('../config')
 
 // Local dependencies
-const extensions = require('./extensions')
+const extensions = require('./plugins')
 const fakeFileSystem = require('../../__tests__/util/mockFileSystem')
 
 // Local variables
