+++ lib/routes/plugins.js
@@ -1,5 +1,5 @@
 const requests = require('../../index').requests
-const extensions = require('../extensions/extensions')
+const extensions = require('../plugins/plugins')
 
 // Serve assets from extensions
 function setupPathsFor (item) {
