+++ lib/routes/prototype-admin-routes.js
@@ -9,7 +9,7 @@ const redirectingRouter = require('../../index').requests.setupRouter('/manage-p
 
 const { projectDir, packageDir } = require('../path-utils')
 const path = require('path')
-const extensions = require('../extensions/extensions')
+const extensions = require('../plugins/plugins')
 const config = require('../config').getConfig()
 const { requestHttpsJson, prototypeAppScripts } = require('../utils')
 const { exec } = require('../exec')
