+++ lib/utils.js
@@ -11,7 +11,7 @@ const { marked } = require('marked')
 
 // Local dependencies
 const config = require('./config').getConfig()
-const extensions = require('./extensions/extensions')
+const extensions = require('./plugins/plugins')
 const filters = require('./filters/api')
 const routes = require('./routes/api')
 const { appDir, tempDir, projectDir } = require('./path-utils')
